add support for more templating options
scscript_template can be used to name a javascript function that will be called to render each resource. script in config can be used to load external Javascript files. the function named will be based a map containing all the properties of the corresponding resource (keyed with full property URIs and '@' for the resource URI), the DOM node container  as a JQuery, the endpoint URL to fetch further data if needed.